requirements.txt
# ============================================
# Google Cloud Platform
# ============================================
google-cloud-aiplatform>=1.38.0      # Vertex AI SDK (embeddings, LLM, vector search)
google-cloud-storage>=2.14.0          # Cloud Storage client
google-cloud-bigquery>=3.14.0         # BigQuery client
google-cloud-logging>=3.9.0           # Cloud Logging
google-cloud-functions>=1.13.0        # Cloud Functions framework
google-auth>=2.25.0                   # GCP authentication

# ============================================
# Data Processing
# ============================================
pandas>=2.1.0                         # Data manipulation
numpy>=1.26.0                         # Numerical computing
pyarrow>=14.0.0                       # Parquet file support
pillow>=10.1.0                        # Image processing

# ============================================
# Medical Imaging
# ============================================
pydicom>=2.4.0                        # DICOM file handling
SimpleITK>=2.3.0                      # Medical image processing
opencv-python>=4.8.0                  # Computer vision utilities

# ============================================
# Machine Learning
# ============================================
scikit-learn>=1.3.0                   # ML utilities
sentence-transformers>=2.2.0          # Local embedding fallback

# ============================================
# API & Web
# ============================================
fastapi>=0.104.0                      # API framework
uvicorn[standard]>=0.24.0             # ASGI server
pydantic>=2.5.0                       # Data validation
pydantic-settings>=2.1.0              # Settings management
httpx>=0.25.0                         # Async HTTP client

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0                  # Environment variable loading
tqdm>=4.66.0                          # Progress bars
tenacity>=8.2.0                       # Retry logic
structlog>=23.2.0                     # Structured logging

# ============================================
# Development & Testing
# ============================================
pytest>=7.4.0                         # Testing framework
pytest-asyncio>=0.21.0                # Async test support
pytest-cov>=4.1.0                     # Coverage reporting
black>=23.11.0                        # Code formatting
ruff>=0.1.6                           # Linting
mypy>=1.7.0                           # Type checking
ipykernel>=6.27.0                     # Jupyter kernel

# ============================================
# Documentation
# ============================================
mkdocs>=1.5.0                         # Documentation generator
mkdocs-material>=9.5.0                # Material theme
